fix bug on message app
