--- src/app/direct/inbox/@message/default.tsx
@@ -1,8 +1,0 @@
-export default async function Default({
-  params,
-}: {
-  params: Promise<{ artist: string }>;
-}) {
-  const artist = (await params).artist;
-  return <div>Click Message</div>;
-}
